erp/erp/requirements.txt
appdirs==1.4.4
APScheduler==3.8.1
asgiref==3.5.0
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.11
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-filter==2.3.0
django-grappelli==2.14.2
django-oauth-toolkit==1.3.2
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djongo==1.3.3
drf-yasg==1.17.1
et-xmlfile==1.1.0
idna==2.10
inflection==0.5.1
iniconfig==1.0.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
mariadb==1.0.9
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mod-wsgi==4.9.0
more-itertools==8.4.0
oauthlib==3.1.0
openpyxl==3.0.9
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
psycopg2==2.8.5
psycopg2-binary==2.9.3
py==1.9.0
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.10.0
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
regex==2020.7.14
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.2.4
toml==0.10.1
typed-ast==1.4.1
tzdata==2021.5
tzlocal==4.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
XlsxWriter==3.0.2
